requirements.txt
openai
langchain
streamlit
faiss-cpu
tiktoken
langchain-community
langchain-openai
python-dotenv